chore: move auth data handling to repository and organize logic
