+++ src/service/AuthService.ts
@@ -2,7 +2,6 @@ import * as jwt from 'jsonwebtoken';
 import * as bcrypt from 'bcrypt';
 import { Tutor } from '../entity/Tutor';
 import { Student } from '../entity/Student';
-import { MysqlDataSource } from '../config/database';
 
 export class AuthService {
   static generateToken(id: number, email: string, role: string): string {
@@ -23,14 +22,5 @@ export class AuthService {
       email: string;
       role: string;
     };
-  }
-
-  static async findUserByEmail(email: string) {
-    const tutorRepository = MysqlDataSource.getRepository(Tutor);
-    const studentRepository = MysqlDataSource.getRepository(Student);
-    const user =
-      (await tutorRepository.findOne({ where: { email } })) ||
-      (await studentRepository.findOne({ where: { email } }));
-    return user;
   }
 }
